Converted calls API over to context bridge, removed some unnecessary logging
