+++ webpack.config.main.js
@@ -16,8 +16,7 @@ module.exports = merge(base, {
     entry: {
         index: './src/main/app/index.ts',
         desktopAPI: './src/main/preload/desktopAPI.js',
-        preload: './src/main/preload/mattermost.js',
-        callsWidget: './src/main/preload/callsWidget.js',
+        preload: './src/main/preload/externalAPI.js',
     },
     externals: {
         'macos-notification-state': 'require("macos-notification-state")',
